Implement Interface for VASP
